requirements.txt
click>=6.7
Flask>=0.12.2
itsdangerous>=0.24
Jinja2>=2.10.1
MarkupSafe>=1.0
Werkzeug>=0.12.2
speedtest-cli>=2.1.2
